ImageGenerator/requirements.txt
torch>=2.1.0
numpy>=1.24.0
Pillow>=10.0.0
diffusers>=0.27.0
transformers>=4.36.0
accelerate>=0.27.0
pyyaml>=6.0.0
safetensors>=0.4.0
sentencepiece>=0.1.99
bitsandbytes>=0.43.0
peft>=0.10.0
openai>=1.6.0
google-generativeai>=0.5.0
tqdm>=4.66.0

